% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkScreenCSV.R
\name{checkScreenCSV}
\alias{checkScreenCSV}
\title{Check the provided CSV file}
\usage{
checkScreenCSV(x, name = "Compound_ID")
}
\arguments{
\item{x}{Path to the csv file}

\item{name}{The expected column name. Defaults to 'Compound_ID'}
}
\value{
Multiple messages may be produced depending on the input.
Eventually returns \code{invisible(logical(1))}.
}
\description{
Check the provided CSV file for existence and the Compound_ID
column
}
\details{
This function initially checks for the existence of the file.
Once confirmed, the first line is loaded and the requested column either
detected, or not.
}
\examples{
fl <- system.file("extdata", "test.csv", package = "dsExtra")
checkScreenCSV(fl, "Col1")
}
