% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictions.R
\name{getPredictions}
\alias{getPredictions}
\title{Get Predictions For Each Test Compound}
\usage{
getPredictions(
  compound,
  fit,
  data,
  column = "Label",
  dist = c("mahalanobis.dist", "mahalanobis.dist", "max.dist")
)
}
\arguments{
\item{compound}{Character vector of compounds on which predictions are to be
performed}

\item{fit}{The object returned when fitting the training dataset using
\code{link{splsda}}}

\item{data}{Data as imported with all measurements used in the training
model}

\item{column}{The column in \code{data} which identifies the compounds}

\item{dist}{The distance measure to use when classifying compounds}
}
\value{
A \code{link{tibble}} containing all original metadata along with
predictions in the first few columns
}
\description{
Get Best Predictions For Each Test Compound
}
\details{
After training a model based on the selected reference compounds, obtain the
predictions for the remaining test compounds.
The best distance measure can be chosen, and concordance across both
components is provided in the output
}
