% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPalette.R
\name{plotPalette}
\alias{plotPalette}
\title{Plot a colour palette}
\usage{
plotPalette(pal, alpha = TRUE)
}
\arguments{
\item{pal}{Character vector of colours. Can be in any recosgnised R format}

\item{alpha}{Include transparency in the colours}
}
\value{
A ggplot object showing the colours provided in the palette
}
\description{
Plot a colour palette as a strip plot
}
\details{
Takes a named vector of colours and produces a single row plot showing each
name and the corresponding colour
}
\examples{
x <- c(A = "blue", B = "yellow", C = "green")
plotPalette(x)

}
