% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParamSummary.R
\name{plotParamSummary}
\alias{plotParamSummary}
\title{Produce a Summary Plot Of All Parameters}
\usage{
plotParamSummary(
  x,
  sep = "_",
  .into = c("Stat", "Target", "Feature", "Value"),
  barCol = "grey70",
  cirCol = rgb(0, 0.5, 0.5, 0.2),
  rng = c(5, 25),
  nudge_top = 0.15,
  nudge_labs = 20,
  ...
)
}
\arguments{
\item{x}{Character vector of colnames from the data used for training the
model}

\item{sep}{The common separator for the parameters}

\item{.into}{The column names after separating the parameters into a df}

\item{barCol}{Colour of the marginal barplots drawn at top & right}

\item{cirCol}{The colour for the circles}

\item{rng}{The size range for the plotted circles}

\item{nudge_top}{Value between 0 & 0.5. Lower values are better.
This can be tweaked to ensure the top marginal plot alings the y-axis with
the y-axis of the lower plot. As these tend to shift with resizing the plot
this may take manual tweaking.}

\item{nudge_labs}{Nudge the labels away from the bars in the marginal
barplots}

\item{...}{Passed to \code{\link{scale_size}}}
}
\value{
A cowplot object composed of four panels
}
\description{
Produce a summary plot of all parameters used for training
the model
}
\details{
Given a character vector of parameeters used for fitting the model, this
functions assumes they have a common separator and are listed in a
hierarchical manner.
A summary plot of all measured parameters, along with marginal values is
produced
}
\examples{
mat <- cbind("Median", rep(c("Membrane", "Nucleus"), each = 5),
rep(c("Colour", "Texture"), each = 10), rep(letters[1:10], times = 2))
param <- apply(mat[1:18,], MARGIN = 1, FUN = paste, collapse = "_")
plotParamSummary(param, nudge_top = 0.07)

}
