% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRefCompounds.R
\name{plotRefCompounds}
\alias{plotRefCompounds}
\title{Plot the Reference Compounds}
\usage{
plotRefCompounds(fit, pal, shapes, size = 1)
}
\arguments{
\item{fit}{The object returned when fitting the training dataset using
\code{link{splsda}}}

\item{pal}{The colour palette for each reference compound}

\item{shapes}{The shapes for each reference compound. (Optional)}

\item{size}{Size of points on the final plot}
}
\description{
Plot the reference compounds after training the model
}
\details{
Takes a trained model with the reference compounds only and plots the PLS
fit as a gglot2 object
}
