% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showNamedCompounds.R
\name{showNamedCompounds}
\alias{showNamedCompounds}
\title{Show all Non-Numeric values in a column}
\usage{
showNamedCompounds(df, col = "Compound_ID")
}
\arguments{
\item{df}{The data frame with named compounds}

\item{col}{The column name to show values from}
}
\value{
Sorted, unique values from the requested column as a character vector.
Only values which cannot be transformed into a numeric value are returned.
}
\description{
Show all Non-Numeric values in a column
}
\details{
Looks in a column, converts everything to a numeric and returns the values
which could not be converted to a numeric value.
This is particualrly useful as many screening compounds are simply given
a numeric identifier. Better known compounds are generally named.

This is useful just for quickly checking values when constructing
abbreviated labels for common compounds.
}
\examples{
df <- data.frame(Compound_ID = c(1:10, "DHT"))
showNamedCompounds(df)

}
